chat2db/requirements.txt
aiofiles == 23.1.0
asgi-correlation-id == 4.3.1
fastapi==0.110.2
fastapi-pagination==0.12.19
itsdangerous==2.1.2
langchain==0.1.16
langchain-openai == 0.0.8
langchain-community == 0.0.34
langchain-text-splitters==0.0.1
more-itertools==10.1.0
psycopg2-binary==2.9.9
python-multipart==0.0.6
python-docx==1.1.0
pandas==2.2.2
pgvector==0.2.5
pydantic==1.10.13
python-dotenv == 1.0.0
requests == 2.31.0
SQLAlchemy == 2.0.23
uvicorn==0.21.0
urllib3==2.2.1
nltk==3.8.1
